Extracts components from the 'App.tsx' file
