+++ resources/src/App.tsx
@@ -1,8 +1,10 @@
 import * as React from "react";
-import {Code, GeistProvider, Link, Loading, Modal, Note, Page, Row, Text, useModal} from "@geist-ui/react";
-import {Info} from "@geist-ui/react-icons";
-import {DataFetcher, ErrorMessageContext} from "./Data";
+import {GeistProvider, Page, Text} from "@geist-ui/react";
+import {DataFetcher} from "./Data";
 import {DataTable} from "./Table";
+import {AppLoader} from "./AppLoader";
+import {ErrorMessage} from "./ErrorMessage";
+import {AppFooter} from "./AppFooter";
 import 'inter-ui/inter.css'
 import './App.css'
 
@@ -22,38 +24,5 @@ const App = () =>
             </Page.Footer>
         </Page>
     </GeistProvider>;
-
-const AppLoader = () =>
-    <Row style={{padding: '10px 0'}}>
-        <Loading>Loading</Loading>
-    </Row>;
-
-const ErrorMessage = () => {
-    const error = React.useContext(ErrorMessageContext);
-    const {setVisible, bindings} = useModal();
-
-    return <>
-        <Note type='error'>
-            Can't fetch the translation progress data from GitHub
-            <span style={{verticalAlign: 'middle', cursor: 'pointer'}}>
-                {' '} <Info size={16} onClick={() => setVisible(true)}/>
-            </span>
-        </Note>
-        <Modal {...bindings} width='50rem'>
-            <Modal.Title>Error</Modal.Title>
-            <Modal.Subtitle>Can't fetch the data from GitHub</Modal.Subtitle>
-            <Modal.Content style={{overflow: 'auto'}}>
-                <Code block width='100%'>{error}</Code>
-            </Modal.Content>
-        </Modal>
-    </>;
-}
-
-const AppFooter = () => {
-    return <Text p>
-        Generated by <Link href="https://github.com/LukWebsForge/TldrProgress" icon> tldr-translation-progress</Link>
-        • Thanks for using this site
-    </Text>;
-}
 
 export {App}
