+++ resources/src/AppFooter.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+import {Link, Text} from "@geist-ui/react";
+
+const AppFooter = () => {
+    return <Text p>
+        Generated by <Link href="https://github.com/LukWebsForge/TldrProgress" icon> tldr-translation-progress</Link>
+        • Thanks for using this site
+    </Text>;
+}
+
+export {AppFooter}
