+++ resources/src/AppLoader.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+import {Loading, Row} from "@geist-ui/react";
+
+const AppLoader = () =>
+    <Row style={{padding: '10px 0'}}>
+        <Loading>Loading</Loading>
+    </Row>;
+
+export {AppLoader}
